more fixes for 0.6
