+++ src/OhMyREPL.jl
@@ -109,7 +109,6 @@ function __init__()
             main_mode.keymap_dict = Base.LineEdit.keymap([d, main_mode.keymap_dict])
         end
     end
-    if false
     mktemp() do _, f
         old_stderr = STDERR
         redirect_stderr(f)
@@ -118,20 +117,8 @@ function __init__()
         if VERSION > v"0.5-" && VERSION.minor < 6
             include(joinpath(dirname(@__FILE__), "errormessage_overrides.jl"))
         end
-
-        @eval begin
-            function Base.REPL.display(d::Base.REPL.REPLDisplay, mime::MIME"text/plain", x)
-                global OUTPUT_PROMPT
-                io = Base.REPL.outstream(d.repl)
-                write(io, OUTPUT_PROMPT)
-                Base.have_color && write(io, Base.REPL.answer_color(d.repl))
-                show(IOContext(io, :limit => true), mime, x)
-                println(io)
-            end
-        end
-
+        include(joinpath(dirname(@__FILE__), "output_prompt_overwrite.jl"))
         redirect_stderr(old_stderr)
-    end
     end
 end
